+++ src/features/search/ui/handlers/searchFormSubmitHandler.ts
@@ -1,12 +1,8 @@
-import { addMoviePost } from "../../../../shared/ui/renderers/addMoviePost";
-import { showSkeletons } from "../../../../shared/ui/renderers/showSkeletons";
-import { getSearchedPost } from "../../api/getSearchedPost";
-import { disableMoreButton } from "../../../../shared/ui/renderers/disabledMoreButton";
-import { IMovie } from "../../../../shared/types/movies";
 import { showErrorPage } from "../../../../shared/ui/renderers/showErrorPage";
 import { setParams } from "../../../../shared/utils/setParams";
 import { pageManager } from "../../../../shared/domain/pageManager";
-import { disableHeaderImage } from "../../../../shared/ui/renderers/disableHeaderImage";
+import { updateSearchPageUI } from "../renderers/updateSearchPageUI";
+import { getSearchedPost } from "../../api/getSearchedPost";
 
 export const searchFormSubmitHandler = async (e: Event) => {
   try {
@@ -27,27 +23,3 @@ export const searchFormSubmitHandler = async (e: Event) => {
     showErrorPage();
   }
 };
-
-export function updateSearchPageUI(
-  searchedMovies: IMovie[],
-  searchQuery: string,
-  { pageNum, totalPages }: { pageNum: number; totalPages: number }
-) {
-  const $thumbnailList = document.querySelector(
-    ".thumbnail-list"
-  ) as HTMLElement;
-  const $movieListTitle = document.querySelector(".movie-list-title");
-
-  if (!$thumbnailList || !$movieListTitle) return;
-
-  $thumbnailList.innerHTML = "";
-  showSkeletons($thumbnailList);
-
-  $movieListTitle.textContent = `"${searchQuery}" 검색 결과`;
-
-  $thumbnailList.innerHTML = "";
-  addMoviePost(searchedMovies, $thumbnailList);
-
-  disableHeaderImage();
-  disableMoreButton(totalPages, pageNum, searchedMovies);
-}
